requirements.txt
langchain
google-generativeai
langchain-google-genai>=1.0.0
sentence-transformers>=2.2.2
torch>=2.0.0
qdrant-client>=1.7.0
chromadb>=0.4.0
python-dotenv>=1.0.0
loguru>=0.7.0
pyYAML>=6.0
langchain-community>=0.0.20
numpy>=1.24.0
scipy>=1.10.0